fix: renamed playerHealth.js
