+++ src/phaser/player/playerStat.js
